+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.17
+
+### Added
+
+- `generate` now supports `parallelGenerationThreads` option. Positive values will turn on parallel generation, and
+  the value will be used as the number of threads. The default is 0, which means no parallel generation.
+
 ## 1.16
 
 ### Added